refactor: 자바 스크립트 변환
