+++ DUNOESANCHAEG-Client/jsconfig.json
@@ -3,9 +3,8 @@
     "target": "ESNext",
     "lib": ["ESNext", "DOM", "DOM.Iterable"],
     "module": "ESNext",
-    /* 🎯 Bundler로 설정하면 Vite 환경에서 모듈을 더 잘 찾습니다. */
     "moduleResolution": "Bundler",
-    "strict": true,
+    "strict": false,
     "jsx": "preserve",
     "sourceMap": true,
     "resolveJsonModule": true,
@@ -15,9 +14,13 @@
     "paths": {
       "@/*": ["src/*"]
     },
-    /* 🎯 핵심: Vite 환경 변수와 클라이언트 타입을 인식시킵니다. */
     "types": ["vite/client"]
   },
-  /* 🎯 d.ts 파일도 포함시켜야 타입 정의가 프로젝트 전체에 적용됩니다. */
-  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue"]
+  "include": [
+    "src/**/*.ts",
+    "src/**/*.js",
+    "src/**/*.d.ts",
+    "src/**/*.tsx",
+    "src/**/*.vue"
+  ]
 }
