+++ DUNOESANCHAEG-Client/src/api/instance.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { useAuthStore } from '@/stores/auth';
+import { useAuthStore } from '@/stores/auth.js'; // 🎯 .js 확장자 명시 확인!
 
 const instance = axios.create({
     baseURL: import.meta.env.VITE_API_BASE_URL,
@@ -9,9 +9,12 @@ const instance = axios.create({
 instance.interceptors.request.use((config) => {
     const authStore = useAuthStore();
     if (authStore.accessToken) {
-        config.headers.Authorization = `Bearer ${authStore.accessToken}`; //
+        config.headers.Authorization = `Bearer ${authStore.accessToken}`;
     }
+
     return config;
+}, (error) => {
+    return Promise.reject(error);
 });
 
 export default instance;
