+++ DUNOESANCHAEG-Client/src/pages/login/KakaoCallback.vue
@@ -17,9 +17,9 @@ import { Loading as VanLoading, showToast } from 'vant';
 import { onMounted } from 'vue';
 import { useRoute, useRouter } from 'vue-router';
 import type { AxiosError } from 'axios'; // 🎯 Axios 에러 타입 임포트
-import instance from '@/api/instance';
-import { useAuthStore } from '@/stores/auth';
-import { getRoleFromToken } from '@/utils/jwtUtils';
+import instance from '@/api/instance.js';
+import { useAuthStore } from '@/stores/auth.js';
+import { getRoleFromToken } from '@/utils/jwtUtils.js';
 
 interface ApiErrorResponse {
   status: number;
